fix: Move Direction enum into next text package
